Extract shared media update rules for native tooling
